WARNING: UNTESTED
Not having the run configurations available for testing, this update is untested.  However, it should currently have no impact on the behaviour of run configurations.

Changes:

Window Range/Limit settings: Join Bolts (Jena and SQL) now have many constructors, allowing for the customisation of range and limit of each of the left and right windows independently.

Bolt output stream flexibility: toFire is now a map of stream name to Values.  Also, fire() and emit() are now overloaded to allow specification of the stream to emit on.  All existing calls to fire and emit remain unchanged in syntax and effect, firing and emitting on the default stream as before.  ReteStormQueues now receive overflow from the CircularPriorityQueue they wrap, and emit the overflowed tuples on non-default streams named after the bolt from which the window's input is streamed.  This is not used at the moment, but could be used in node level optimisation.