Latest changes:
switch from boolean "to build or not to build, so to probe" to an enum stored in the abstract StormGraphRouter, comprising the list of actions that could be intended by any given query-processing message within the system.  This includes a full propagation of subsequent changes required by this original change.