BingImageDataset#get(int) now returns null for bad image.
Before, the BingImageDataset used to throw a RunTimeException
if the http response was not an image. This could happen if an
image URL redirected to a .html page (such as in tinypic.com URLs).

For example, using Bing search for "dog" or "cat" would come up
with some, and would crash the program.

In addition, the code now throws an IOException if an HTTP 401 Error is
recieved, which happens if the API Key is incorrect.