Adds DefaultTokenFactory#delete(Class<T>) convenience function
The DefaultTokenFactory already had convenience functions to get
a cached token/create a new one. However, if one entered the incorrect
API Key, or generated new API Keys and invalidated previous ones, it
was difficult to delete these. This change allows a user to see
a delete function on Eclipse's autocomplete.